finally adding typedef

+++ include/cjks/io.h
@@ -42,13 +42,13 @@ CJKS_DLL int cjks_io_read(cjks_io *io, void *buf, size_t len);
 CJKS_DLL int cjks_io_write(cjks_io *io, const void *buf, size_t len);
 CJKS_DLL int cjks_io_close(cjks_io *io);
 
-CJKS_DLL unsigned short cjks_io_read_be2(cjks_io *io);
-CJKS_DLL unsigned int cjks_io_read_be4(cjks_io *io);
-CJKS_DLL unsigned long long cjks_io_read_be8(cjks_io *io);
+CJKS_DLL uint16 cjks_io_read_be2(cjks_io *io);
+CJKS_DLL uint32 cjks_io_read_be4(cjks_io *io);
+CJKS_DLL uint64 cjks_io_read_be8(cjks_io *io);
 
-CJKS_DLL int cjks_io_write_be2(cjks_io *io, unsigned short s);
-CJKS_DLL int cjks_io_write_be4(cjks_io *io, unsigned int i);
-CJKS_DLL int cjks_io_write_be8(cjks_io *io, unsigned long long l);
+CJKS_DLL int cjks_io_write_be2(cjks_io *io, uint16 s);
+CJKS_DLL int cjks_io_write_be4(cjks_io *io, uint32 i);
+CJKS_DLL int cjks_io_write_be8(cjks_io *io, uint64 l);
 
 CJKS_DLL char *cjks_io_aread_utf(cjks_io *io);
 CJKS_DLL int cjks_io_aread_data(cjks_io* io, cjks_buf *buf);
